[refactor] 리셀 인사이트 공용 모델 및 좌석 등급 API 분리
